Added new scn drone and modified the classes. See PR for changes
