+++ ShootingApp/Resources/texts/drones.txt
@@ -1,0 +1,2 @@
+"Animated Drone" (https://skfb.ly/6WOAz) by hartwelkisaka is licensed under Creative Commons Attribution (http://creativecommons.org/licenses/by/4.0/).
+
